IS-3330: Sett sykmeldt som ufordelt
